User denis

+++ src/main/java/gr36/clubActiv/controller/ActivityController.java
@@ -8,6 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -36,17 +37,13 @@ public class ActivityController {
   }
 
   @PostMapping
-  public ResponseEntity<?> create(@RequestBody ActivityDto activity) {
-    try {
-      log.info("Creating new activity: {}", activity);
-      ActivityDto createdActivity = service.create(activity);
-      log.info("Activity created successfully with ID: {}", createdActivity.getId());
-      return ResponseEntity.status(HttpStatus.CREATED).body(createdActivity);
-    } catch (Exception e) {
-      log.error("Error while creating activity", e);
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to create activity");
-    }
+  public ResponseEntity<ActivityDto> create(@RequestBody ActivityDto activity) {
+    log.info("Creating new activity: {}", activity);
+    ActivityDto createdActivity = service.create(activity);
+    log.info("Activity created successfully with ID: {}", createdActivity.getId());
+    return ResponseEntity.status(HttpStatus.CREATED).body(createdActivity);
   }
+
 
   @GetMapping
   public List<ActivityDto> getAllActivities() {
@@ -56,52 +53,46 @@ public class ActivityController {
 
   @GetMapping("/{id}")
   public ResponseEntity<ActivityDto> getActivityById(@PathVariable Long id) {
-    try {
       log.info("Fetching activity by ID: {}", id);
       ActivityDto activity = service.getActivityById(id);
       return ResponseEntity.ok(activity);
-    } catch (EntityNotFoundException e) {
-      log.error("Activity not found with ID: {}", id, e);
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
-    } catch (Exception e) {
-      log.error("Error while fetching activity by ID: {}", id, e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
   }
 
   @PutMapping("/update/{id}")
   public ResponseEntity<?> update(
       @PathVariable("id") Long id,
       @RequestBody ActivityDto dto) {
-    try {
       log.info("Updating activity with ID: {}", id);
       ActivityDto updatedActivity = service.update(id, dto);
       return ResponseEntity.ok(updatedActivity);
-    } catch (EntityNotFoundException e) {
-      log.error("Activity not found with ID: {}", id, e);
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Activity not found");
-    } catch (Exception e) {
-      log.error("Error while updating activity with ID: {}", id, e);
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to update activity");
-    }
   }
+
+//  @DeleteMapping("/{id}")
+//  public ResponseEntity<?> deleteActivity(@PathVariable Long id) {
+//      log.info("Deleting activity with ID: {}", id);
+//      service.deleteActivity(id);
+//      return ResponseEntity.noContent().build();
+//   }
 
   @DeleteMapping("/{id}")
+  @PreAuthorize("hasRole('ADMIN')")
   public ResponseEntity<?> deleteActivity(@PathVariable Long id) {
-    try {
-      log.info("Deleting activity with ID: {}", id);
-      service.deleteActivity(id);
-      return ResponseEntity.noContent().build();
-    } catch (EntityNotFoundException e) {
-      log.error("Activity not found with ID: {}", id, e);
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Activity not found");
-    } catch (Exception e) {
-      log.error("Error while deleting activity with ID: {}", id, e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to delete activity");
+    log.info("Attempting to delete activity with ID: {}", id);
+
+    // Проверяем, существует ли активность
+    ActivityDto activity = service.getActivityById(id);
+    if (activity == null) {
+      log.error("Activity with ID {} not found", id);
+      return new ResponseEntity<>("Activity not found", HttpStatus.NOT_FOUND);
     }
+
+    // Если активность существует, удаляем ее
+    service.deleteActivity(id);
+    log.info("Activity with ID {} deleted successfully", id);
+    return ResponseEntity.noContent().build();
   }
 
-  @PostMapping("/{activity_id}/add-user/{user_id}")
+  @PutMapping("/{activity_id}/add-user/{user_id}") //TODO POST -> PUT
   public ResponseEntity<ActivityDto> addUserToActivity(@PathVariable Long activity_id, @PathVariable Long user_id) {
     ActivityDto updatedActivity = service.addUserToActivity(activity_id, user_id);
     return ResponseEntity.ok(updatedActivity); // Возвращаем обновленную активность
@@ -112,5 +103,12 @@ public class ActivityController {
     log.info("Fetching activities for user ID: {}", userId); // Логирование
     return service.getActivitiesByUserId(userId);
   }
-}
 
+  @DeleteMapping("/{activity_id}/remove-user/{user_id}")
+  public ResponseEntity<?> removeUserFromActivity(@PathVariable Long activity_id, @PathVariable Long user_id) {
+    log.info("Delete activities for user ID: {}", user_id); // Логирование
+      ActivityDto updatedActivity = service.removeUserFromActivity(activity_id, user_id);
+      return ResponseEntity.ok(updatedActivity);
+  }
+
+}
